Formated the code and added the linters

+++ routes/api.php
@@ -8,10 +8,7 @@ Route::post('/login', 'UserController@login')->name('login');
 Route::get('get_site_settings', 'SettingsController@get_site_settings');
 
 Route::get('test', function () {
-    $js = \App\Http\Controllers\EmployeesController::sendAccountCreationEmail(
-        'irfanullah@techleadz.com',
-        'John'
-    );
+    $js = \App\Http\Controllers\EmployeesController::sendAccountCreationEmail('irfanullah@techleadz.com', 'John');
     return $js;
 });
 
@@ -25,7 +22,7 @@ Route::group(['middleware' => 'auth:api'], function () {
     Route::post('site_configuration', 'SettingsController@site_configuration');
     Route::get('testTbl', 'SettingsController@testTbl');
 
-    Route::group(["prefix" => 'v1'], function () {
+    Route::group(['prefix' => 'v1'], function () {
         Route::get('profile', 'UserController@profile');
         Route::post('profile', 'UserController@update');
 
@@ -35,18 +32,9 @@ Route::group(['middleware' => 'auth:api'], function () {
         Route::apiResource('employees', 'EmployeesController');
 
         Route::get('employees_all', 'EmployeesController@employees_all');
-        Route::get(
-            'permission_by_group',
-            'PermissionsController@permission_by_group'
-        );
-        Route::post(
-            'roles_assign_permissions',
-            'RolesController@roles_assign_permissions'
-        );
-        Route::get(
-            'permission_by_role/{role}',
-            'PermissionsController@permission_by_role'
-        );
+        Route::get('permission_by_group', 'PermissionsController@permission_by_group');
+        Route::post('roles_assign_permissions', 'RolesController@roles_assign_permissions');
+        Route::get('permission_by_role/{role}', 'PermissionsController@permission_by_role');
         //Employe Assign Roles
         Route::get('get_user_role/{user}', 'EmployeesController@get_user_role');
         Route::post('assign_role', 'EmployeesController@assign_role');
@@ -55,10 +43,7 @@ Route::group(['middleware' => 'auth:api'], function () {
         Route::post('permissions', 'PermissionsController@store');
         Route::get('permissions/{permission}', 'PermissionsController@show');
         Route::put('permissions/{permission}', 'PermissionsController@update');
-        Route::delete(
-            'permissions/{permission}',
-            'PermissionsController@destroy'
-        );
+        Route::delete('permissions/{permission}', 'PermissionsController@destroy');
 
         Route::apiResource('students', 'StudentsController');
